% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REQAO.R
\name{AchieveLabel}
\alias{AchieveLabel}
\title{AchieveLabel(df, grade, type) (char/num) - Apply Labels to Achievement Levels EQAO 3 6 9 and choose type of label}
\usage{
AchieveLabel(x, grade, type)
}
\description{
Use:   AchLabel(df, grade, type)
}
\details{
grade: 3 (recodes ROverallLevel, WOVerallLevel and MOverallLevel)
grade: 6 (recodes R...W...MOverallLevel, Prior_G3_R...Prior_G3_W...Prior_G3_MOverallLevel)
grade: 9 (recodes OverallOutcomeLevel, Prior_G6_R...W...M, Prior_G3_R...W...MOverallLevel)
grade: 10 (recodes OSSLTOutcome, Prior G6_R...W, Prior_G3_R...WOverallLevel)

type: char = character labels (Level 1, Level 2, Witheld etc.)
type: num = numeric (NE1 = 0, all non-levels are NA - like fully participating)

- pass a variable and recode with labels for achievement levels
}
