% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REQAO.R
\name{ProvLoad}
\alias{ProvLoad}
\title{df <- ProvLoad(3, YYYY, datadir)}
\usage{
ProvLoad(grade, year, datadir)
}
\description{
Use:   df <- SchoolLoad(3, 2017, datadir)
grades = 3, 6, 9, 10
year = YYYY (school year in June)
}
\details{
- identify the grade, year: used to construct EQAO file names
- for the DDSB use v and c to access network and c data directories
  otherwise pass a character string or vector with the directory containing the EQAO files
}
