% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REQAO.R
\name{SchoolLoad}
\alias{SchoolLoad}
\title{Function to Load EQAO School Files  ------------------------------------}
\usage{
SchoolLoad(grade, year, bident, board, datadir)
}
\description{
Use:   df <- SchoolLoad(3, 2017, 66060, "DDSB", datadir)
grades = 3, 6, 9, 10
year = YYYY (school year in June)
}
\details{
- identify the grade, year and board id: used to construct EQAO file names
- identify the school board: used to label the board row in the file which is NA by default
- for the DDSB use v and c to access network and c data directories
  otherwise pass a character string or vector with the directory containing the EQAO files
}
