% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REQAO.R
\name{StudentLoad}
\alias{StudentLoad}
\title{Examples
test3 <- SchoolLoad(3, 2017, 66060, "DDSB")
test6 <- SchoolLoad(6, 2017, 66060, "DDSB")
test9 <- SchoolLoad(9, 2017, 66060, "DDSB")
test10 <- SchoolLoad(10, 2017, 66060, "DDSB")
testfail <- SchoolLoad(8, 2017, 6606, "DDSB")
Function to Load Student Files  ------------------------------------}
\usage{
StudentLoad(grade, year, bident, datadir)
}
\description{
Use:   df <- StudentLoad(3, 2017, 66060, datadir)
grades = 3, 6, 9, 10
}
\details{
- identify the grade, year and board id: used to construct EQAO file names
- for the DDSB use v and c to access network and c data directories
  otherwise pass a character string or vector with the directory containing the EQAO files
}
