% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REQAO.R
\name{AchieveLabel}
\alias{AchieveLabel}
\title{AchieveLabel(df$var, grade, type) (char/num) - Apply Labels to Achievement Levels EQAO 3 6 9 and choose type of label}
\usage{
AchieveLabel(x, grade, type)
}
\description{
Use:   AchLabel(df$var, type)
type: char = character labels (Level 1, Level 2, Witheld etc.)
type: num = numeric (NE1 = 0, all non-levels are NA - like fully participating)
}
\details{
- pass a variable and recode with labels for achievement levels
}
